Remove move command
It is impossible to detect file moves by just looking at the contents.
There needs to be some lower level primitive to uniquely identify a file
at the time of creation (UUID for instance) to be able to detect file
move. Tested with watchdog and watchgod but they also do not provide a
file move event. It is always translated to delete+add event.